msm: cpr-regulator: add support to choose vdd_apc from multiple sources
Cpr-regulator needs to detect several possible regulator supplies because
there is no way to tell boards apart that have each of the possible
regulators, the difference can't be captured in device tree. Modify the
cpr-regulator driver so that it chooses the first available probed
regulator as its vdd_apc source.

Change-Id: I7c7fda08dcd8bad884720b9550576f7d7ad5220b
Signed-off-by: Ke Liu <keliu@codeaurora.org>

Conflicts:
	Documentation/devicetree/bindings/arm/msm/cpr-regulator.txt